requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-multipart
python-dotenv
pandas
openpyxl
python-jose[cryptography]
passlib[bcrypt]
